- Implemented cogs to seperates functions     + Fftcg.py carries all of the querying functions     + Management.py carries all of the bot management functions     + Marcie.py carries all of the functions for bot events
- Created helper functions, which clean up `name` and `image` a little bit
    + cardlistToEmbed
    + cardToNameEmbed
    + cardToImageEmbed
    + selectLogic

- Introduced `paginate` which utilizes DiscordUtils to allow for paginating through cards in UI
