+++ constants.py
@@ -1,0 +1,2 @@
+EMBEDCOLOR = 0xd93fb6
+MAX_QUERY = 35
